backend/api/routers/auth.py
"""
Authentication endpoints
"""
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from pydantic import BaseModel, EmailStr
from typing import Optional
from datetime import datetime, timedelta
from jose import JWTError, jwt
from passlib.context import CryptContext

router = APIRouter()

# Password hashing
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth2
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")


class Token(BaseModel):
    access_token: str
    token_type: str


class TokenData(BaseModel):
    email: Optional[str] = None


class UserCreate(BaseModel):
    email: EmailStr
    password: str
    full_name: Optional[str] = None


class UserResponse(BaseModel):
    id: int
    email: str
    full_name: Optional[str]
    is_active: bool

    class Config:
        from_attributes = True


@router.post("/register", response_model=UserResponse)
async def register(user: UserCreate):
    """Register a new user"""
    # TODO: Implement user registration
    return {
        "id": 1,
        "email": user.email,
        "full_name": user.full_name,
        "is_active": True
    }


@router.post("/token", response_model=Token)
async def login(form_data: OAuth2PasswordRequestForm = Depends()):
    """Login and get access token"""
    # TODO: Implement authentication
    return {
        "access_token": "dummy_token",
        "token_type": "bearer"
    }


@router.get("/me", response_model=UserResponse)
async def get_current_user(token: str = Depends(oauth2_scheme)):
    """Get current user"""
    # TODO: Implement user retrieval from token
    return {
        "id": 1,
        "email": "user@example.com",
        "full_name": "Test User",
        "is_active": True
    }


@router.get("/garmin/authorize")
async def garmin_authorize():
    """Initiate Garmin OAuth flow"""
    # TODO: Implement Garmin OAuth
    return {"authorization_url": "https://connect.garmin.com/oauth"}


@router.get("/garmin/callback")
async def garmin_callback(code: str):
    """Handle Garmin OAuth callback"""
    # TODO: Implement OAuth callback handling
    return {"status": "success"}


@router.get("/oura/authorize")
async def oura_authorize():
    """Initiate Oura OAuth flow"""
    # TODO: Implement Oura OAuth
    return {"authorization_url": "https://cloud.ouraring.com/oauth"}


@router.get("/oura/callback")
async def oura_callback(code: str):
    """Handle Oura OAuth callback"""
    # TODO: Implement OAuth callback handling
    return {"status": "success"}


@router.get("/strava/authorize")
async def strava_authorize():
    """Initiate Strava OAuth flow"""
    # TODO: Implement Strava OAuth
    return {"authorization_url": "https://www.strava.com/oauth/authorize"}


@router.get("/strava/callback")
async def strava_callback(code: str):
    """Handle Strava OAuth callback"""
    # TODO: Implement OAuth callback handling
    return {"status": "success"}
