backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# API Clients
garminconnect==0.2.5
stravalib==1.4.0

# Configuration
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Data Validation
email-validator==2.1.0

# Utilities
tenacity==8.2.3
python-slugify==8.0.1
